refactor: move functions for getting puzzle name to own module puzzle_name.js
