+++ nonograms/scripts/helpers/puzzle_name.js
@@ -1,0 +1,14 @@
+import * as vars from '../variables.js';
+
+function getPuzzleNameLevel(isTable) {
+  const regex = /\(.*\)/g;
+  const puzzleName = vars.pictures[vars.currentLevelId].name
+    .replace(regex, '').trim();
+  const puzzleLevel = vars.pictures[vars.currentLevelId].level
+    .replace(regex, '').trim();
+  return isTable ? `${puzzleName} - ${puzzleLevel}` : [puzzleName, puzzleLevel];
+}
+
+export {
+  getPuzzleNameLevel
+};
